added photoUrl
