+++ src/AllApi/MainApi.jsx
@@ -1,2 +1,2 @@
-export const ServerApi = 'https://find-job-server.vercel.app'
-// export const ServerApi = "http://localhost:5001";
+// export const ServerApi = 'https://find-job-server.vercel.app'
+export const ServerApi = "http://localhost:5001";
