+++ src/Component/Login/Login.jsx
@@ -32,7 +32,7 @@ const Login = () => {
                 // console.log(result)
                 if (result.user) {
                     const user = result.user;
-                    if (!user.emailVerified === true) {
+                    if (user.emailVerified === true) {
                         toast.error('Please Verify You email.....!')
                         userVerification()
                             .then(() => {
